backend/pyproject.toml
[project]
name = "simulationcity-backend"
version = "0.1.0"
description = "SimulationCity backend — cooperative real-time city building game"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.111",
    "python-socketio>=5.11",
    "uvicorn[standard]>=0.29",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "beanie>=1.26",
    "motor>=3.4",
    "pymongo>=4.7",
    "celery[redis]>=5.4",
    "redis[hiredis]>=5.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
premium = [
    "replicate>=0.29",
    "boto3>=1.34",
    "Pillow>=10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "locust>=2.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "workers"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
